Make python pre-commit hook fail if it would reformat
Currently, when I run `git commit` with a staged file that would be
reformatted by the pre-commit hook, the hook will reformat the file but
the the commit succeed with the unforamtted staged files.

I'd like to make the pre-commit hook fail if it would reformat the file,
so that I can fix the formatting and then commit.