systemd controller manipulation (#240)
Add code to manipulate the Systemd cgroups hierarchy:
On Cgroups V1, manipulate the name=systemd controller.
On Cgroups V2, manipulate the cgroup hierarchy regardless of supported controllers (empty controller).

This is needed because Systemd keeps track of process assignment to units by their cgroups: if the V2 hierarchy is mounted, it acts as the source of truth - regardless of which controllers are enabled or delegated. If the V2 hierarchy is not mounted, a name=systemd controller is created to the V1 hierarchy and become the source of truth.